chore: upgrade deps, migrate to TypeScript, and fix security issues
- Convert worker.js and theme/index.js to TypeScript with tsconfig

- Upgrade wrangler and add pnpm overrides to resolve 16 audit findings

- Pin GitHub Actions to latest stable versions (checkout@v7, setup-node@v6, etc.)
